removed junk and implemented new event system
