--- engineFoxtrot/Engine/Events/EventDispatcher.h
@@ -1,24 +1,0 @@
-#pragma once
-#include "Events/Event.h"
-
-class EventDispatcher
-{
-public:
-	EventDispatcher(Event& event)
-		: m_Event(event)
-	{
-	}
-
-	template<typename T, typename F>
-	bool Dispatch(const F& func)
-	{
-		if (m_Event.GetEventType() == T::GetStaticType())
-		{
-			m_Event.Handled = func(static_cast<T&>(m_Event));
-			return true;
-		}
-		return false;
-	}
-private:
-	Event& m_Event;
-};
