+++ engineFoxtrot/Engine/Events/Mouse/MouseScrolledEvent.h
@@ -1,0 +1,21 @@
+#pragma once
+#include "Events/Event.h"
+
+class MouseScrolledEvent : public Event
+{
+public:
+	MouseScrolledEvent(const float xOffset, const float yOffset)
+		: m_XOffset(xOffset), m_YOffset(yOffset) {}
+
+	float GetXOffset() const { return m_XOffset; }
+	float GetYOffset() const { return m_YOffset; }
+
+	std::string ToString() const override
+	{
+		std::stringstream ss;
+		ss << "MouseScrolledEvent: " << GetXOffset() << ", " << GetYOffset();
+		return ss.str();
+	}	
+private:
+	float m_XOffset, m_YOffset;
+};
