+++ engineFoxtrot/Main/main.cpp
@@ -3,7 +3,6 @@
 #include "../Engine/Engine.h"
 
 #include "Events/Window/WindowCloseEvent.h"
-#include "Events/EventDispatcher.h"
 
 // TODO engine.h & engine.cpp
 
@@ -17,9 +16,6 @@ Engine engine;
 
 void sceneTestSetup()
 {
-	engine.addEventListener(new PhysicsEngine, OldEventType::MOVE);
-	engine.eventManager.notify(OldEventType::MOVE, new Object(1));
-
 	engine.createNewSceneWithSceneID(3);
 
 	engine.linkSpriteIDWithAssetPath(1, "../Assets/Sprites/Project assets/Adventurer-1.5/Individual Sprites/adventurer-air-attack1-00.png");
